+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Latest
 
+- Fixed APILoader so it now saves duration, distance, etc. for all protocols.
+
 ## 0.3.1 (2025-07-24)
 
 - Updated APILoader as familyComName and familySciName in the data from eBird, is optional.